integrate transfer-patient feature into patient-detail page
