+++ src/app/app.module.ts
@@ -10,7 +10,6 @@ import { PatientPage } from '../pages/patient/patient';
 import { AddPatientPage } from '../pages/add-patient/add-patient';
 import { PatientDetailPage } from '../pages/patient-detail/patient-detail';
 import { UpdatePatientPage } from '../pages/update-patient/update-patient';
-import { TransferPatientModal } from '../pages/transfer-patient/transfer-patient';
 import { NurseListPage } from '../pages/nurse-list/nurse-list';
 import { DoctorListPage } from '../pages/doctor-list/doctor-list';
 import { NurseProvider } from '../providers/nurse/nurse';
@@ -34,7 +33,6 @@ import { LoginProvider } from '../providers/login/login';
     AddPatientPage,
     PatientDetailPage,
     UpdatePatientPage,
-    TransferPatientModal,
     AdminTabs,
     NurseListPage,
     DoctorListPage,
@@ -56,7 +54,6 @@ import { LoginProvider } from '../providers/login/login';
     AddPatientPage,
     PatientDetailPage,
     UpdatePatientPage,
-    TransferPatientModal,
     AdminTabs,
     NurseListPage,
     DoctorListPage,
