--- src/pages/transfer-patient/transfer-patient.scss
@@ -1,2 +1,0 @@
-transfer-patient {
-}
